Some cleanup in preparation for the next release
Completed EDGRTAC-4 and EDGRTAC-5.

Cleaned up the POM file a bit to remove some warnings and add consistency to vertx dependencies. The vertx version remains the same (updating to 3.5.4 caused some problems with tests that will need to be looked into).

Added an async on the `MainVerticleTest` shutdown so it will complete within the test context. However, the mock Okapi server will still not complete its shutdown within the context of the test. This is addressed in edge-common 1.0.1. This is also true for `RtacOkapiClientTest` which makes use of the mock Okapi server.

The `ephemeral.properties` had to be updated to support the modified format required by 1.0.0 edge-common.